+++ solana/Anchor.toml
@@ -17,7 +17,7 @@ xfluencer = "DmYaabL1PhacWWsRwyZpBqBP9n7tVq7115zG2tznYLb9"
 url = "https://api.apr.dev"
 
 [provider]
-cluster = "localnet"
+cluster = "mainnet"
 wallet = "/home/ruben/.config/solana/id.json"
 
 [scripts]
