+++ solana/programs/xfluencer/src/processor/mod.rs
@@ -1,5 +1,5 @@
 pub mod initialize_escrow;
-pub mod create_escrow_solana;
+pub mod create_escrow_sol;
 pub mod cancel;
 pub mod claim;
 pub mod claim_escrow;
